Added DE2 target
This sort-of looks like it's working (the address shows on the 7-segment displays) but I don't have a picture on my monitor. It could be something like positive vs. negative sync signals or something. I'll try another monitor later and I'll also add code to switch to serial (this appears to be hardwired off).